Moved SeqSelectOption to interact.py
